+++ src/Pages/ABL.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { Table } from "react-bootstrap";
 import { useHistory } from "react-router-dom";
 import moment from "moment";
-import Navbar from "../Navbar/Navbar";
+import NavbarComponent from "../components/navbar/navbar.component";
 import { sortTable } from "../utils/TableUtils";
 
 const ABL = () => {
@@ -99,7 +99,7 @@ const ABL = () => {
     <div>
       {ablItems ? (
         <>
-          <Navbar />
+          <NavbarComponent />
           <section>
             <Table hover>
               <thead>
@@ -261,7 +261,7 @@ const ABL = () => {
         </>
       ) : (
         <>
-          <Navbar />
+          <NavbarComponent />
           <section></section>
         </>
       )}
